mod usize to u64 for consistency

+++ anchor/programs/befundr/src/constants.rs
@@ -1,6 +1,6 @@
 pub mod user {
-    pub const MAX_NAME_LENGTH: usize = 64;
-    pub const MAX_AVATAR_URL_LENGTH: usize = 128;
-    pub const MAX_BIO_LENGTH: usize = 256;
-    pub const MAX_CITY_LENGTH: usize = 100;
+    pub const MAX_NAME_LENGTH: u64 = 64;
+    pub const MAX_AVATAR_URL_LENGTH: u64 = 128;
+    pub const MAX_BIO_LENGTH: u64 = 256;
+    pub const MAX_CITY_LENGTH: u64 = 100;
 }
